Fix a softlock when inhaled by Gyorg and dying simultaneously
Fixes a possible softlock when dying while being inhaled by Gyorg.

Prior to this fix, the game would softlock because that sequence of
actions causes Link to exit the "held (by Gyorg or Majora's whips)"
state and eventually enter (and stay in) the frozen state after
the fairy revival cutscene.