Fix softlock when hitting owl statue after playing ocarina
Clearing the player's pending cutscene ID is normally done by z_player
before it calls the talk function, but we *must* do it manually before
talking to the owl statue because we are calling the talk function
directly without going through z_player.

Failure to clear the pending cutscene ID can have catastrophic results
because the owl actor triggers a cutscene; we get a nasty deadlock if
the player code also attempts to trigger a cutscene.

The easiest way to trigger this issue was to play the ocarina and then
immediately strike an inactive owl statue.

Fixes #167