Fix unsafe player access in UpdatePadState
We should *not* access the global context or the player actor manually
in UpdatePadState because neither are guaranteed to still exist in
memory when that function is called.

Fixes random crashes if ZR is held during a transition (transitions
usually involve destructing the current game state and then
reinitialising again -- the global context may be reallocated at a
different location). Closes #146.